CC10 - Height Weight Screen

+++ core/src/main/java/com/example/core/data/preferences/DefaultPreferences.kt
@@ -28,9 +28,9 @@ class DefaultPreferences(
             .apply()
     }
 
-    override fun saveHeight(weight: Float) {
+    override fun saveHeight(weight: Int) {
         sharedPref.edit()
-            .putFloat(Preferences.KEY_WEIGHT, weight)
+            .putInt(Preferences.KEY_WEIGHT, weight)
             .apply()
     }
